src/sn_stage.c: Fix comparison of signed and unsigned integers.

+++ src/sn_stage.c
@@ -431,7 +431,7 @@ int sn_stage_serialize (sn_stage_t *s,
 
 #define SNPRINTF_OR_FAIL(...) \
   status = snprintf (buffer, buffer_size, __VA_ARGS__); \
-  if ((status < 1) || (status >= buffer_size)) \
+  if ((status < 1) || (((size_t) status) >= buffer_size)) \
     return (-1); \
   buffer += status; \
   buffer_size -= status;
